pyproject.toml
[project]
name = "mangrag"
version = "1.0.0"
description = "Production RAG pipeline — MongoDB Atlas + Groq + HuggingFace"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dnspython>=2.7.0",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.115.0",
    "groq>=0.13.0",
    "httpx>=0.27.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "pymongo>=4.15.5",
    "pypdf>=6.4.1",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.40.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
mangrag = "main:main"
mangrag-api = "uvicorn api:app --reload"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
